personalized_notifications/requirements.txt
pandas
numpy
requests
geopy
jupyter
tensorflow
gym
pyspark
kafka-python
apache-airflow
starburst
snowflake-connector-python
